Updates for ubuntu:latest, python 3.10, werkzeug 2.3.7, remove meinheld (#31)
* Updates for ubuntu:latest, python 3.10, werkzeug 2.3.7

Due to 441 reported CVEs in this image (Ubuntu focal packages and
python libraries), I found it necessary to update the base OS to
ubuntu:latest. Doing so brings in Python 3.10 and updated pipenv. To
resolve the python library CVEs, update the dependencies which include
werkzeug 2.3.7 which necessitates a few minor changes to the
application.

* remove meinheld

The meinheld library is no longer needed to run gunicorn, and it has
a medium CVE reported against it. Upgrading gunicorn to latest and
removing meinheld library will resolve this.

* Use ubuntu:jammy-20230816