issue 62 fix (#63)
Adding the HOME variable makes the Docker build process create and use the /httpbin/.local folder in both build and runtime. Without this, the Dockerfile creates a folder /root/.local and uses it while the specified user is root. However, in the case of a user override with runAsUser or a similar command, /root is no longer the home directory and is not accessible to the non-root user. This line fixes this issue. It has been fixed and tested locally.